refactor(hooks): move use-marquee-selection into shared/hooks
Used by 4 features plus the shared marquee overlay, so it belongs in
shared/ alongside other cross-feature primitives. Removes the lone
src/hooks/ folder.
