+++ src/shared/hooks/use-marquee-selection.ts
